+++ kwcoco/__init__.py
@@ -269,7 +269,7 @@ Testing:
 
 """
 
-__version__ = '0.5.6'
+__version__ = '0.5.7'
 
 
 __submodules__ = {