+++ CHANGELOG.md
@@ -5,7 +5,22 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.5.6 - Unreleased
+## Version 0.5.7 - Unreleased
+
+### Changed
+* In `metrics.drawing`, make threshold colors more consistent
+* Added `category_names` property to `Annots1d` object.
+
+
+### Added
+* Added `annots` property to CocoImge
+
+### Fixed 
+* Better asset space support
+* Querying `annots` by track-id now return them in a sorted order.
+
+
+## Version 0.5.6 - Released 2023-02-07
 
 ### Added
 * Can now ask for the resolution of an asset in CocoImage.